backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
bluetooth-numbers
bleak
scapy
python-dotenv
